Various fixes to ships and buy from vendor procedure.
Fixed: Ships moving but not updating the multi part on Enhanced Client with smooth sailing disabled.
Changed: Removed useless screen updates when smooth sailing is on, making sailing more fluid.
Changed: Removed declarations and calculations on unused variables in the sailing engine.
Fixed: Price of items sold by the vendor swapped on Enhanced Client.
Changed: Removed unnecessary casts and changed variable type from INT64 to LONG in CItemVendable::GetVendorPrice and CClient::Event_VendorSell.
Changed: Removed useless packets sent when opening vendor sell gump on Enhanced Client.

+++ src/graysvr/CItemVend.cpp
@@ -166,7 +166,7 @@ LONG CItemVendable::GetVendorPrice( int iConvertFactor )
 	//    0 = base price
 	// +100 = increase price by 100% (vendor selling to player?)
 
-	INT64 lPrice = m_price;
+	LONG lPrice = m_price;
 	if ( lPrice <= 0 )	// set on player vendor.
 	{
 		if ( lPrice == 0 )	// set a new randomized price for the item
@@ -184,7 +184,7 @@ LONG CItemVendable::GetVendorPrice( int iConvertFactor )
 				pItemDef = Item_GetDef();
 			}
 			lPrice = pItemDef->GetMakeValue(GetQuality());
-			m_price = static_cast<long>(-lPrice);
+			m_price = -lPrice;
 		}
 		else
 		{
@@ -198,7 +198,7 @@ LONG CItemVendable::GetVendorPrice( int iConvertFactor )
 	else if (lPrice <= 0)
 		return 0;
 	
-	return static_cast<long>(lPrice);
+	return lPrice;
 }
 
 bool CItemVendable::IsValidSaleItem( bool fBuyFromVendor ) const
